program looks in /var/log/auth.log for evil.
This is a little python program that looks in /var/log/auth.log (or a copy that you made, probably not a good to run this as root), looking for login failures.  When it finds one, it attempts to geolocate the origin by IP address.

It has not been tested with IPv6.